backend/requirements.txt
# Core Django packages
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1

# API Documentation
drf-yasg==1.21.7

# Utilities
django-filter==23.5
django-extensions==3.2.3
Pillow==10.2.0

# Celery for async tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Email
django-anymail==10.2

# Security
django-ratelimit==4.1.0
django-guardian==2.4.0

# Monitoring and Logging
sentry-sdk==1.39.2
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Django storage
django-storages==1.14.2
boto3==1.34.14

# Caching
django-redis==5.4.0

# Data validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27